- properties for files could be committed and fetched via propget - class DirEntry is defined from the properties (and all dir entry properties are stored as properties and not as instance variable) - dates in direntry are jSvnServe timestamps (so that they could handle microseconds)

+++ trunk/src/main/java/com/googlecode/jsvnserve/api/editorcommands/AbstractDeltaFile.java
@@ -194,7 +194,9 @@ throw new ServerException("temporary file '" + this.file + "' not found");
             } catch (final IOException e) {
             }
         }
-        this.file.delete();
+        if (this.file != null)  {
+            this.file.delete();
+        }
     }
 
 }
